+++ tests/app/e2e/steps/about-dialog.js
@@ -1,5 +1,5 @@
 // @flow
-import { Given, When, Then } from 'cucumber';
+import { Given, When, Then } from '@cucumber/cucumber';
 import { expect } from 'chai';
 import packageJson from '../../../../package.json';
 import type { Daedalus } from '../../../types';