+++ tests/app/e2e/steps/node-restart.js
@@ -1,5 +1,5 @@
 // @flow
-import { Given, When, Then } from 'cucumber';
+import { Given, When, Then } from '@cucumber/cucumber';
 import { waitUntilTextInSelector } from '../../../common/e2e/steps/helpers';
 import type { Daedalus } from '../../../types';