+++ tests/paper-wallets/e2e/steps/paper-wallets-certificate.js
@@ -1,5 +1,5 @@
 // @flow
-import { Given, When, Then } from 'cucumber';
+import { Given, When, Then } from '@cucumber/cucumber';
 import { expect } from 'chai';
 import path from 'path';
 import { fillOutWalletSendForm } from '../../../wallets/e2e/steps/helpers';